Refactoring to prepare for Collections of rows of restaurants.
