+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/home/HomeScreen.kt
@@ -15,10 +15,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import androidx.lifecycle.viewmodel.compose.viewModel
 import androidx.navigation.NavController
 import androidx.navigation.compose.rememberNavController
-import com.jordangellatly.hermosahappyhour.model.RestaurantViewModel
+import com.jordangellatly.hermosahappyhour.model.sampleSearchRestaurantData
 import com.jordangellatly.hermosahappyhour.ui.components.BottomNavigationBar
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
 import java.text.SimpleDateFormat
@@ -28,7 +27,6 @@ import java.util.*
 fun HomeScreen(
     navController: NavController
 ) {
-    val restaurantViewModel: RestaurantViewModel = viewModel()
     Scaffold(
         bottomBar = {
             BottomNavigationBar(navController = navController)
@@ -80,7 +78,7 @@ fun HomeScreen(
                 horizontalArrangement = Arrangement.spacedBy(20.dp),
                 modifier = Modifier.padding(start = 16.dp, end = 16.dp)
             ) {
-                itemsIndexed(restaurantViewModel.sampleSearchRestaurantData) { index, restaurant ->
+                itemsIndexed(sampleSearchRestaurantData) { index, restaurant ->
                     HighlightRestaurantItem(
                         restaurant = restaurant,
                         onRestaurantClick = {
@@ -107,7 +105,7 @@ fun HomeScreen(
                 horizontalArrangement = Arrangement.spacedBy(20.dp),
                 modifier = Modifier.padding(start = 16.dp, end = 16.dp)
             ) {
-                itemsIndexed(restaurantViewModel.sampleSearchRestaurantData) { index, restaurant ->
+                itemsIndexed(sampleSearchRestaurantData) { index, restaurant ->
                     HighlightRestaurantItem(
                         restaurant = restaurant,
                         onRestaurantClick = {
@@ -134,7 +132,7 @@ fun HomeScreen(
                 horizontalArrangement = Arrangement.spacedBy(20.dp),
                 modifier = Modifier.padding(start = 16.dp, end = 16.dp)
             ) {
-                itemsIndexed(restaurantViewModel.sampleSearchRestaurantData) { index, restaurant ->
+                itemsIndexed(sampleSearchRestaurantData) { index, restaurant ->
                     HighlightRestaurantItem(
                         restaurant = restaurant,
                         onRestaurantClick = {
@@ -161,7 +159,7 @@ fun HomeScreen(
                 horizontalArrangement = Arrangement.spacedBy(20.dp),
                 modifier = Modifier.padding(start = 16.dp, end = 16.dp)
             ) {
-                itemsIndexed(restaurantViewModel.sampleSearchRestaurantData) { index, restaurant ->
+                itemsIndexed(sampleSearchRestaurantData) { index, restaurant ->
                     HighlightRestaurantItem(
                         restaurant = restaurant,
                         onRestaurantClick = {
