+++ app/src/main/java/com/jordangellatly/hermosahappyhour/ui/home/RestaurantCard.kt
@@ -4,9 +4,16 @@ import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.*
 import androidx.compose.foundation.lazy.LazyRow
 import androidx.compose.foundation.lazy.items
+import androidx.compose.foundation.lazy.itemsIndexed
+import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material.Icon
+import androidx.compose.material.IconButton
 import androidx.compose.material.MaterialTheme
 import androidx.compose.material.Text
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.outlined.ArrowBack
+import androidx.compose.material.icons.outlined.ArrowForward
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -16,16 +23,21 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.text.style.TextOverflow
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.Dp
 import androidx.compose.ui.unit.dp
 import coil.compose.AsyncImage
 import coil.request.ImageRequest
 import com.jordangellatly.hermosahappyhour.R
+import com.jordangellatly.hermosahappyhour.model.CollectionType
 import com.jordangellatly.hermosahappyhour.model.Restaurant
+import com.jordangellatly.hermosahappyhour.model.RestaurantCollection
+import com.jordangellatly.hermosahappyhour.model.sampleSearchRestaurantData
 import com.jordangellatly.hermosahappyhour.ui.components.HappyHourCard
 import com.jordangellatly.hermosahappyhour.ui.components.HermosaHappyHourSurface
 import com.jordangellatly.hermosahappyhour.ui.components.offsetGradientBackground
 import com.jordangellatly.hermosahappyhour.ui.theme.HermosaHappyHourTheme
+import com.jordangellatly.hermosahappyhour.ui.utils.mirroringIcon
 
 val HighlightCardWidth = 170.dp
 val HighlightCardPadding = 16.dp
@@ -36,6 +48,87 @@ private val gradientWidth
     get() = with(LocalDensity.current) {
         (3 * (HighlightCardWidth + HighlightCardPadding).toPx())
     }
+
+@Composable
+fun RestaurantCollection(
+    restaurantCollection: RestaurantCollection,
+    onRestaurantClick: (Long) -> Unit,
+    modifier: Modifier = Modifier,
+    index: Int = 0,
+    highlight: Boolean = true
+) {
+    Column(modifier = modifier) {
+        Row(
+            verticalAlignment = Alignment.CenterVertically,
+            modifier = Modifier
+                .heightIn(min = 56.dp)
+                .padding(start = 24.dp)
+        ) {
+            Text(
+                text = restaurantCollection.name,
+                style = MaterialTheme.typography.h6,
+                color = HermosaHappyHourTheme.colors.brand,
+                maxLines = 1,
+                overflow = TextOverflow.Ellipsis,
+                modifier = Modifier
+                    .weight(1f)
+                    .wrapContentWidth(Alignment.Start)
+            )
+            IconButton(
+                onClick = { /* todo */ },
+                modifier = Modifier.align(Alignment.CenterVertically)
+            ) {
+                Icon(
+                    imageVector = mirroringIcon(
+                        ltrIcon = Icons.Outlined.ArrowForward,
+                        rtlIcon = Icons.Outlined.ArrowBack
+                    ),
+                    tint = HermosaHappyHourTheme.colors.brand,
+                    contentDescription = null
+                )
+            }
+        }
+        if (highlight && restaurantCollection.type == CollectionType.Highlight) {
+            HighlightedRestaurants(index, restaurantCollection.snacks, onRestaurantClick)
+        } else {
+            Restaurants(restaurantCollection.snacks, onRestaurantClick)
+        }
+    }
+}
+
+@Composable
+private fun HighlightedRestaurants(
+    index: Int,
+    restaurants: List<Restaurant>,
+    onRestaurantClick: (Long) -> Unit,
+    modifier: Modifier = Modifier
+) {
+    val scroll = rememberScrollState(0)
+    val gradient = when ((index / 2) % 2) {
+        0 -> HermosaHappyHourTheme.colors.gradient6_1
+        else -> HermosaHappyHourTheme.colors.gradient6_2
+    }
+    // The Cards show a gradient which spans 3 cards and scrolls with parallax.
+    val gradientWidth = with(LocalDensity.current) {
+        (6 * (HighlightCardWidth + HighlightCardPadding).toPx())
+    }
+    LazyRow(
+        modifier = modifier,
+        horizontalArrangement = Arrangement.spacedBy(16.dp),
+        contentPadding = PaddingValues(start = 24.dp, end = 24.dp)
+    ) {
+        itemsIndexed(restaurants) { index, snack ->
+            HighlightRestaurantItem(
+                snack,
+                onRestaurantClick,
+                index,
+                gradient,
+                gradientWidth,
+                scroll.value
+            )
+        }
+    }
+}
 
 @Composable
 private fun Restaurants(
@@ -182,6 +275,35 @@ fun RestaurantImage(
             placeholder = painterResource(R.drawable.placeholder),
             modifier = Modifier.fillMaxSize(),
             contentScale = ContentScale.Crop,
+        )
+    }
+}
+
+@Preview
+@Composable
+fun RestaurantImagePreview() {
+    HermosaHappyHourTheme {
+        RestaurantImage(
+            imageUrl = "\"https://scontent-lax3-2.xx.fbcdn.net/v/t39.30808-6/292336365_435159251951538_4078078271979326231_n.jpg?_nc_cat=107&ccb=1-7&_nc_sid=09cbfe&_nc_ohc=C_tAC12x4o0AX9uN-Yg&_nc_ht=scontent-lax3-2.xx&oh=00_AfAAxhYzzBx-bK8gDhRHmOeMrzUxDDsDMZeCaTyFHRr3Ug&oe=638F052F\"",
+            contentDescription = null,
+            modifier = Modifier
+                .size(120.dp)
+        )
+    }
+}
+
+@Preview
+@Composable
+fun RestaurantCardPreview() {
+    HermosaHappyHourTheme {
+        val restaurant = sampleSearchRestaurantData.first()
+        HighlightRestaurantItem(
+            restaurant = restaurant,
+            onRestaurantClick = {},
+            index = 0,
+            gradient = HermosaHappyHourTheme.colors.gradient6_1,
+            gradientWidth = gradientWidth,
+            scroll = 0
         )
     }
 }
